Make matrix-project dependency optional
